reorg code a bit
