+++ frontend/src/p2/__tests__/ConfirmationSummary.test.js
@@ -1,0 +1,35 @@
+import React from 'react'
+import { render, cleanup } from '@testing-library/react'
+import { ApolloProvider } from 'react-apollo'
+import { ThemeProvider } from 'emotion-theming'
+import { I18nProvider } from '@lingui/react'
+import { ConfirmationSummary } from '../ConfirmationSummary'
+import theme from '../../theme'
+import en from '../../../locale/en/messages.js'
+
+const client = {
+  readQuery: () => ({
+    scamInfo: JSON.stringify({}),
+    lostMoney: JSON.stringify({}),
+    suspectInfo: JSON.stringify({}),
+    files: [],
+    contactInfo: JSON.stringify({}),
+  }),
+  writeData: jest.fn(),
+}
+
+describe('<ConfirmationSummary />', () => {
+  afterEach(cleanup)
+
+  it('renders', () => {
+    render(
+      <ThemeProvider theme={theme}>
+        <ApolloProvider client={client}>
+          <I18nProvider language={'en'} catalogs={{ en }}>
+            <ConfirmationSummary />
+          </I18nProvider>
+        </ApolloProvider>
+      </ThemeProvider>,
+    )
+  })
+})
